+++ packages/jest-core/src/__tests__/SearchSource.test.ts
@@ -40,7 +40,7 @@ const toPaths = (tests: Array<Test>) => tests.map(({path}) => path);
 let findMatchingTests: (config: Config.ProjectConfig) => Promise<SearchResult>;
 
 describe('SearchSource', () => {
-  const name = 'SearchSource';
+  const id = 'SearchSource';
   let searchSource: SearchSource;
 
   describe('isTestFilePath', () => {
@@ -49,7 +49,7 @@ describe('SearchSource', () => {
     beforeEach(() => {
       config = normalize(
         {
-          name,
+          id,
           rootDir: '.',
           roots: [],
         },
@@ -68,7 +68,7 @@ describe('SearchSource', () => {
       if (process.platform !== 'win32') {
         config = normalize(
           {
-            name,
+            id,
             rootDir: '.',
             roots: [],
             testMatch: undefined,
@@ -117,8 +117,8 @@ describe('SearchSource', () => {
     it('finds tests matching a pattern via testRegex', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx', 'txt'],
-          name,
           rootDir,
           testMatch: undefined,
           testRegex: 'not-really-a-test',
@@ -141,8 +141,8 @@ describe('SearchSource', () => {
     it('finds tests matching a pattern via testMatch', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx', 'txt'],
-          name,
           rootDir,
           testMatch: ['**/not-really-a-test.txt', '!**/do-not-match-me.txt'],
           testRegex: '',
@@ -165,8 +165,8 @@ describe('SearchSource', () => {
     it('finds tests matching a JS regex pattern', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx'],
-          name,
           rootDir,
           testMatch: undefined,
           testRegex: 'test.jsx?',
@@ -187,8 +187,8 @@ describe('SearchSource', () => {
     it('finds tests matching a JS glob pattern', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx'],
-          name,
           rootDir,
           testMatch: ['**/test.js?(x)'],
           testRegex: '',
@@ -209,8 +209,8 @@ describe('SearchSource', () => {
     it('finds tests matching a JS with overriding glob patterns', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx'],
-          name,
           rootDir,
           testMatch: [
             '**/*.js?(x)',
@@ -237,7 +237,7 @@ describe('SearchSource', () => {
     it('finds tests with default file extensions using testRegex', () => {
       const {options: config} = normalize(
         {
-          name,
+          id,
           rootDir,
           testMatch: undefined,
           testRegex,
@@ -258,7 +258,7 @@ describe('SearchSource', () => {
     it('finds tests with default file extensions using testMatch', () => {
       const {options: config} = normalize(
         {
-          name,
+          id,
           rootDir,
           testMatch,
           testRegex: '',
@@ -279,7 +279,7 @@ describe('SearchSource', () => {
     it('finds tests with parentheses in their rootDir when using testMatch', () => {
       const {options: config} = normalize(
         {
-          name,
+          id,
           rootDir: path.resolve(__dirname, 'test_root_with_(parentheses)'),
           testMatch: ['<rootDir>**/__testtests__/**/*'],
           testRegex: undefined,
@@ -299,8 +299,8 @@ describe('SearchSource', () => {
     it('finds tests with similar but custom file extensions', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx'],
-          name,
           rootDir,
           testMatch,
         },
@@ -320,8 +320,8 @@ describe('SearchSource', () => {
     it('finds tests with totally custom foobar file extensions', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'foobar'],
-          name,
           rootDir,
           testMatch,
         },
@@ -341,8 +341,8 @@ describe('SearchSource', () => {
     it('finds tests with many kinds of file extensions', () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx'],
-          name,
           rootDir,
           testMatch,
         },
@@ -362,7 +362,7 @@ describe('SearchSource', () => {
     it('finds tests using a regex only', () => {
       const {options: config} = normalize(
         {
-          name,
+          id,
           rootDir,
           testMatch: undefined,
           testRegex,
@@ -383,7 +383,7 @@ describe('SearchSource', () => {
     it('finds tests using a glob only', () => {
       const {options: config} = normalize(
         {
-          name,
+          id,
           rootDir,
           testMatch,
           testRegex: '',
@@ -430,7 +430,7 @@ describe('SearchSource', () => {
               'haste_impl.js',
             ),
           },
-          name: 'SearchSource-findRelatedTests-tests',
+          id: 'SearchSource-findRelatedTests-tests',
           rootDir,
         },
         {} as Config.Argv,
@@ -479,8 +479,8 @@ describe('SearchSource', () => {
     beforeEach(async () => {
       const {options: config} = normalize(
         {
+          id,
           moduleFileExtensions: ['js', 'jsx', 'foobar'],
-          name,
           rootDir,
           testMatch,
         },
@@ -537,7 +537,7 @@ describe('SearchSource', () => {
       if (process.platform !== 'win32') {
         const config = normalize(
           {
-            name,
+            id,
             rootDir: '.',
             roots: ['/foo/bar/prefix'],
           },
@@ -570,7 +570,7 @@ describe('SearchSource', () => {
               '../../../jest-haste-map/src/__tests__/haste_impl.js',
             ),
           },
-          name: 'SearchSource-findRelatedSourcesFromTestsInChangedFiles-tests',
+          id: 'SearchSource-findRelatedSourcesFromTestsInChangedFiles-tests',
           rootDir,
         },
         {} as Config.Argv,