+++ packages/jest-haste-map/src/__tests__/index.test.js
@@ -197,8 +197,8 @@ describe('HasteMap', () => {
     defaultConfig = {
       extensions: ['js', 'json'],
       hasteImplModulePath: require.resolve('./haste_impl.js'),
+      id: 'haste-map-test',
       maxWorkers: 1,
-      name: 'haste-map-test',
       platforms: ['ios', 'android'],
       resetCache: false,
       rootDir: path.join('/', 'project'),
@@ -269,8 +269,8 @@ describe('HasteMap', () => {
   it('creates different cache file paths for different projects', () => {
     jest.resetModules();
     const HasteMap = require('../').default;
-    const hasteMap1 = new HasteMap({...defaultConfig, name: '@scoped/package'});
-    const hasteMap2 = new HasteMap({...defaultConfig, name: '-scoped-package'});
+    const hasteMap1 = new HasteMap({...defaultConfig, id: '@scoped/package'});
+    const hasteMap2 = new HasteMap({...defaultConfig, id: '-scoped-package'});
     expect(hasteMap1.getCacheFilePath()).not.toBe(hasteMap2.getCacheFilePath());
   });