+++ packages/jest-runner/src/index.ts
@@ -156,8 +156,8 @@ export default class TestRunner {
   ) {
     const resolvers: Map<string, SerializableResolver> = new Map();
     for (const test of tests) {
-      if (!resolvers.has(test.context.config.name)) {
-        resolvers.set(test.context.config.name, {
+      if (!resolvers.has(test.context.config.id)) {
+        resolvers.set(test.context.config.id, {
           config: test.context.config,
           serializableModuleMap: test.context.moduleMap.toJSON(),
         });