+++ packages/jest-runner/src/testWorker.ts
@@ -59,9 +59,9 @@ const formatError = (error: string | ErrorWithCode): SerializableError => {
 
 const resolvers = new Map<string, Resolver>();
 const getResolver = (config: Config.ProjectConfig) => {
-  const resolver = resolvers.get(config.name);
+  const resolver = resolvers.get(config.id);
   if (!resolver) {
-    throw new Error('Cannot find resolver for: ' + config.name);
+    throw new Error('Cannot find resolver for: ' + config.id);
   }
   return resolver;
 };
@@ -75,7 +75,7 @@ export function setup(setupData: {
     serializableModuleMap,
   } of setupData.serializableResolvers) {
     const moduleMap = ModuleMap.fromJSON(serializableModuleMap);
-    resolvers.set(config.name, Runtime.createResolver(config, moduleMap));
+    resolvers.set(config.id, Runtime.createResolver(config, moduleMap));
   }
 }