+++ packages/jest-test-sequencer/src/__tests__/test_sequencer.test.js
@@ -23,7 +23,7 @@ const context = {
   config: {
     cache: true,
     cacheDirectory: '/cache',
-    name: 'test',
+    id: 'test',
   },
   hasteFS: {
     getSize: path => path.length,
@@ -34,7 +34,7 @@ const secondContext = {
   config: {
     cache: true,
     cacheDirectory: '/cache2',
-    name: 'test2',
+    id: 'test2',
   },
   hasteFS: {
     getSize: path => path.length,