+++ packages/jest-transform/src/ScriptTransformer.ts
@@ -139,7 +139,7 @@ export default class ScriptTransformer {
   ): Config.Path {
     const baseCacheDir = HasteMap.getCacheFilePath(
       this._config.cacheDirectory,
-      'jest-transform-cache-' + this._config.name,
+      'jest-transform-cache-' + this._config.id,
       VERSION,
     );
     const cacheKey = this._getCacheKey(content, filename, options);