+++ packages/jest-transform/src/__tests__/ScriptTransformer.test.ts
@@ -223,7 +223,7 @@ describe('ScriptTransformer', () => {
     config = makeProjectConfig({
       cache: true,
       cacheDirectory: '/cache/',
-      name: 'test',
+      id: 'test',
       rootDir: '/',
       transformIgnorePatterns: ['/node_modules/'],
     });